unthinkable/requirements.txt
flask
python-dotenv
google-genai
flask-cors